requirements.txt
pandas>=1.3.0
numpy>=1.21.0,<2.0.0
scikit-learn>=1.0.0
scikit-surprise>=1.1.1
matplotlib>=3.4.0
seaborn>=0.11.0
jupyter>=1.0.0
ipykernel>=6.0.0
tqdm>=4.62.0
scipy>=1.7.0